refactor #57: Move commands in folder.
