+++ src/commands/extract.rs
