+++ src/commands/merge.rs
