+++ src/commands/mod.rs
@@ -1,0 +1,3 @@
+pub mod extract;
+pub mod merge;
+pub mod sort;
